fix(local-inference/voice/kokoro): speed-tensor dtype polarity (kokoro-onnx upstream)
The speed-tensor dtype in `KokoroOnnxRuntime.synthesize` was inverted
relative to the `kokoro-onnx` upstream Python (`kokoro_onnx/__init__.py:
_create_audio`). Upstream:

- newer `input_ids` export → `speed: int32 [1]`
- legacy  `tokens` export   → `speed: float32 [1]`

Current code did the opposite, causing every legacy ONNX export (the
INT8 `model_quantized.onnx` at `onnx-community/Kokoro-82M-v1.0-ONNX`,
which uses `tokens`) to fail with:

  Unexpected input data type. Actual: (tensor(int32)), expected: (tensor(float))

End-to-end verified by hitting `POST /api/tts/local-inference` on the
local dev server with the v1.0 INT8 ONNX + `af_bella.bin` staged:
HTTP 200, 183 KB of valid 24 kHz mono WAV, 1.13× RTF through Milady's
runtime chain.

+++ packages/app-core/src/services/local-inference/voice/kokoro/kokoro-runtime.ts
@@ -218,16 +218,17 @@ export class KokoroOnnxRuntime implements KokoroRuntime {
             (Math.min(inputIds.length, numPositions - 1) + 1) * args.voice.dim,
           )
         : fullStyle;
-    // Older `kokoro-onnx` exports name the token input `tokens` and want a
-    // `speed: int32 [1]` scalar; newer exports name it `input_ids` and want
-    // `speed: float32 [1]`. Detect from the loaded session, mirroring the
-    // probe `kokoro-onnx._create_audio` does on `sess.get_inputs()`.
+    // Speed-tensor dtype mirrors `kokoro-onnx` upstream
+    // (`kokoro_onnx/__init__.py:_create_audio`): the legacy `tokens`
+    // export wants `speed: float32 [1]` and the newer `input_ids` export
+    // wants `speed: int32 [1]`. ORT raises `Unexpected input data type`
+    // when either dtype is swapped, so this polarity is load-bearing.
     const inputNames = session.inputNames ?? ["input_ids", "style", "speed"];
     const useLegacyInputs = !inputNames.includes("input_ids");
     const tokensInputName = useLegacyInputs ? "tokens" : "input_ids";
     const speedTensor = useLegacyInputs
-      ? new ort.Tensor("int32", new Int32Array([1]), [1])
-      : new ort.Tensor("float32", new Float32Array([1.0]), [1]);
+      ? new ort.Tensor("float32", new Float32Array([1.0]), [1])
+      : new ort.Tensor("int32", new Int32Array([1]), [1]);
     const feeds: Record<string, OrtTensor> = {
       [tokensInputName]: new ort.Tensor("int64", inputIds, [
         1,
